+++ packages/backend/src/sbom/sbom-utils.ts
@@ -1,71 +1,12 @@
-import axios from "axios";
-import axiosRetry from "axios-retry";
 import { SbomPackage, Severity } from "shared-types";
+import { logger } from "../utils/logger";
 import licenseConfig from "./licenseConfig.json";
-import {
-  Component,
-  Dependency,
-  LicenseInfo,
-  NpmLicense,
-  PackageRequestData,
-} from "./sbom-types";
-// import { PromisePool } from '@supercharge/promise-pool'
+import { isNpmPackage, isPythonPackage } from "./sbom-fetching-utils";
+import { Component, Dependency, LicenseInfo, NpmLicense } from "./sbom-types";
 
 function sortByLicenseLength(arr: LicenseInfo[]): LicenseInfo[] {
   return arr.sort((a, b) => b.license.length - a.license.length);
 }
-
-export const fetchPackages = async (
-  dependencies: Dependency[],
-  components: Component[]
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-): Promise<any[]> => {
-  // filter out deps without "dependsOn"
-  const withDependsOn = dependencies.filter((dep) => !!dep.dependsOn);
-
-  // map all dependsOn values and flat the matrix
-  const flatDeps = withDependsOn.map((dep) => dep.dependsOn).flat();
-
-  // filter out deps without "component"
-  const withComponent = flatDeps.filter((purl) =>
-    components.some((component) => component.purl === purl)
-  );
-
-  const npmPackages = withComponent
-    .filter((purl) => isNpmPackage(purl))
-    .map((purl) => resolvePackageRequestData(purl, components))
-    .filter((pkg) => !!pkg)
-    .map(fetchDataFromNPM);
-
-  const pyPackages = withComponent
-    .filter((purl) => isPythonPackage(purl))
-    .map((purl) => resolvePackageRequestData(purl, components))
-    .filter((pkg) => !!pkg)
-    .map(fetchDataFromPyPi);
-
-  const allPackagesPromises = [...npmPackages, ...pyPackages];
-  const allPackages = await Promise.all(allPackagesPromises);
-  console.log("All packages");
-  console.log(allPackages);
-  return allPackages;
-};
-
-const isPythonPackage = (purl: string) => purl.startsWith("pkg:pypi");
-const isNpmPackage = (purl: string) => purl.startsWith("pkg:npm");
-
-const resolvePackageRequestData = (
-  purl: string,
-  components: Component[]
-): PackageRequestData | undefined => {
-  const component = components.find((component) => component.purl === purl);
-  if (component) {
-    return {
-      name: component.name,
-      version: component.version,
-    };
-  }
-  return undefined;
-};
 
 export async function getPackages(
   dependencies: Dependency[],
@@ -83,11 +24,10 @@ export async function getPackages(
     if (Object.prototype.hasOwnProperty.call(applications, dependency.ref)) {
       filePath = applications[dependency.ref];
     } else {
-      console.log(`no application for ref: ${dependency.ref}`);
+      logger.sbom.log(`no application for ref: ${dependency.ref}`);
     }
     if (dependency.dependsOn) {
       for (const purl of dependency.dependsOn) {
-        console.log(`working on purl ${purl}`);
         const component = components.find(
           (component) => component.purl === purl
         );
@@ -112,7 +52,7 @@ export async function getPackages(
                 sourceList.push(packageInfo.info.classifiers.join(" "));
               }
             } catch (error) {
-              console.error("Error:", error);
+              logger.sbom.error("Error:", error.message);
             }
           } else if (isNpmPackage(purl)) {
             try {
@@ -130,17 +70,17 @@ export async function getPackages(
                   }
                 });
               } else {
-                console.log(`missing license for: ${purl}`);
+                logger.sbom.log(`missing license for: ${purl}`);
               }
             } catch (error) {
-              console.error("Error:", error);
+              logger.sbom.error(error);
             }
           } else {
-            console.log(`purl: ${purl}`);
+            logger.sbom.log(`purl: ${purl}`);
           }
 
           if (sourceList.length == 0) {
-            console.log(`no where to get license for ${purl}`);
+            logger.sbom.log(`no where to get license for ${purl}`);
           } else {
             for (const licenseSoruce of sourceList) {
               const licenseItem = sortedLicenseConfig.find((item) =>
@@ -163,11 +103,11 @@ export async function getPackages(
             filePath: filePath,
           });
         } else {
-          console.log(`missing component info, purl: ${purl}`);
+          logger.sbom.log(`missing component info, purl: ${purl}`);
         }
       }
     } else {
-      console.log(`no depends for ref: ${dependency.ref}`);
+      logger.sbom.log(`no depends for ref: ${dependency.ref}`);
     }
   }
 
@@ -177,36 +117,6 @@ export async function getPackages(
       severityOrder.indexOf(b.severity) - severityOrder.indexOf(a.severity)
   );
   return sbomPackages;
-}
-
-async function fetchDataFromPyPi({
-  name,
-  version,
-}: // eslint-disable-next-line @typescript-eslint/no-explicit-any
-PackageRequestData): Promise<any> {
-  axiosRetry(axios, { retries: 15, retryDelay: axiosRetry.exponentialDelay });
-  try {
-    const url = `https://pypi.org/pypi/${name}/${version}/json`;
-    const response = await axios.get(url);
-    return response.data;
-  } catch (error) {
-    console.error("Error fetching data from PyPi:", error);
-  }
-}
-
-async function fetchDataFromNPM({
-  name,
-  version,
-}: // eslint-disable-next-line @typescript-eslint/no-explicit-any
-PackageRequestData): Promise<any> {
-  axiosRetry(axios, { retries: 15, retryDelay: axiosRetry.exponentialDelay });
-  try {
-    const url = `https://registry.npmjs.org/${name}/${version}`;
-    const response = await axios.get(url);
-    return response.data;
-  } catch (error) {
-    console.error("Error fetching data from NPM:", error);
-  }
 }
 
 export function extractMappingForApplication(