+++ CHANGELOG.md
@@ -23,7 +23,7 @@ _Can be run using `npm run codetotal:beta`_
 - Back-End
   - Bug fix: SBOM packages not showing up in report page. Async parsing of packages information in SBOM module
   - Retry calls to pypi or npm in case first attempts are failing
-
+  - Fetch SBOM packages with concurrency of 10
 ## [v0.5.0] - 2023-08-10
 
 - Global